Move router/CORS/exception helpers into ApiGatewayUtil and wire gateway failure handling
Relocate createRouterWithCors, createExceptionConvertingFailureHandler, and
writeException from kinotic-persistence's VertxWebUtil into the gateway's
ApiGatewayUtil, and have ApiGatewayVertcleFactory build its router via
createRouterWithCors so unhandled failures on the api-gateway routes render as
JSON with a mapped status code (400/401/403/500) instead of Vert.x's default
bare reason-phrase response.

The moved cluster has no persistence coupling — it only touches Vert.x, JDK,
and core exceptions (AuthenticationException/AuthorizationException) — so it
lives cleanly in gateway with no dependency cycle (persistence already depends
on gateway). VertxWebUtil keeps its persistence-specific helpers (validate*,
getPageable*); its callers (GqlVerticle, OpenApiVertxRouterFactory, and the
openapi NoValue/ValueToJson/Count handlers) now call ApiGatewayUtil.
