+++ kinotic-api-gateway/src/main/java/org/kinotic/gateway/api/utils/ApiGatewayUtil.java
@@ -1,19 +1,31 @@
 package org.kinotic.gateway.api.utils;
 
+import io.vertx.core.Handler;
+import io.vertx.core.Vertx;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.http.HttpServerOptions;
+import io.vertx.core.http.HttpServerResponse;
+import io.vertx.core.json.JsonObject;
 import io.vertx.core.net.PemKeyCertOptions;
+import io.vertx.ext.web.Router;
+import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.CorsHandler;
+import lombok.extern.slf4j.Slf4j;
+import org.kinotic.core.api.exceptions.AuthenticationException;
+import org.kinotic.core.api.exceptions.AuthorizationException;
 import org.kinotic.gateway.api.config.CorsProperties;
 import org.kinotic.gateway.api.config.SslProperties;
 
 import java.util.Set;
+import java.util.concurrent.CompletionException;
 
 /**
- * Helpers for configuring the Vert.x HTTP servers fronted by the api-gateway: CORS handling
- * and PEM-based TLS. Applied to every browser-facing route (STOMP, static web, and the
- * openapi/graphql routes mounted by other modules).
+ * Helpers for the Vert.x HTTP servers and routers fronted by the api-gateway: CORS handling,
+ * PEM-based TLS, router assembly, and rendering exceptions as JSON error responses. Applied to
+ * every browser-facing route (STOMP, static web, and the openapi/graphql routes mounted by
+ * other modules).
  */
+@Slf4j
 public final class ApiGatewayUtil {
 
     /**
@@ -62,5 +74,67 @@ public final class ApiGatewayUtil {
                            .setKeyPath(ssl.getKeyPath()));
         }
         return options;
+    }
+
+    /**
+     * Creates a router pre-wired with CORS and the exception-converting failure handler, so
+     * unhandled failures render as JSON with a mapped status code instead of Vert.x's default
+     * bare reason-phrase response.
+     */
+    public static Router createRouterWithCors(Vertx vertx, CorsProperties corsProperties) {
+        Router router = Router.router(vertx);
+        router.route().failureHandler(createExceptionConvertingFailureHandler());
+        router.route().handler(createCorsHandler(corsProperties));
+        return router;
+    }
+
+    public static Handler<RoutingContext> createExceptionConvertingFailureHandler(){
+        return ctx -> {
+            writeException(ctx, ctx.failure());
+        };
+    }
+
+    public static void writeException(RoutingContext context, Throwable throwable){
+        HttpServerResponse response = context.response();
+        String errorMessage;
+        int statusCode = context.statusCode();
+
+        if(throwable != null){
+
+            if(throwable instanceof CompletionException){
+                if(throwable.getCause() != null) {
+                    throwable = throwable.getCause();
+                }
+            }
+
+            errorMessage = throwable.getMessage();
+
+            if(statusCode == -1){
+                if (throwable instanceof IllegalArgumentException) {
+                    statusCode = 400;
+                } else if (throwable instanceof NullPointerException) {
+                    statusCode = 400;
+                } else if(throwable instanceof AuthenticationException){
+                    statusCode = 401;
+                } else if (throwable instanceof AuthorizationException) {
+                    statusCode = 403;
+                } else {
+                    statusCode = 500;
+                }
+            }
+            log.debug("Error processing web request. Status Code ({})", statusCode, throwable);
+        }else{
+
+            if(statusCode == -1){
+                statusCode = 500;
+            }
+            errorMessage = "Server Error";
+
+            log.warn("Unknown exception occurred processing web request. Status Code ({})", statusCode);
+        }
+        String jsonString = new JsonObject().put("error", errorMessage).encode();
+        response.setStatusCode(statusCode);
+        response.putHeader("Content-Type", "application/json");
+        response.end(jsonString);
     }
 }
